# currency-service/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для получения курсов валют с поддержкой gRPC и HTTP",
        "title": "Currency Exchange API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@currency-api.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/rates": {
            "get": {
                "description": "Возвращает курс обмена между двумя валютами на указанную дату",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rates"
                ],
                "summary": "Получить курс валюты",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Базовая валюта (например: USD, RUB, EUR)",
                        "name": "base",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Целевая валюта (например: USD, RUB, EUR)",
                        "name": "target",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата в формате YYYY-MM-DD (например: 2026-02-15)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfascuf_PROD1_currency-service_internal_models.GetRateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfascuf_PROD1_currency-service_internal_models.GetRateResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/rates/history": {
            "get": {
                "description": "Возвращает исторические данные курсов обмена между двумя валютами за указанный период",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rates"
                ],
                "summary": "Получить историю курсов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Базовая валюта (например: USD, RUB, EUR)",
                        "name": "base",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Целевая валюта (например: USD, RUB, EUR)",
                        "name": "target",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Начальная дата в формате YYYY-MM-DD",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Конечная дата в формате YYYY-MM-DD",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfascuf_PROD1_currency-service_internal_models.GetHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfascuf_PROD1_currency-service_internal_models.GetHistoryResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Проверка состояния API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_alfascuf_PROD1_currency-service_internal_models.ExchangeRate": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rate": {
                    "type": "number"
                },
                "target": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_alfascuf_PROD1_currency-service_internal_models.GetHistoryResponse": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_alfascuf_PROD1_currency-service_internal_models.ExchangeRate"
                    }
                },
                "error": {
                    "type": "string"
                },
                "target": {
                    "type": "string"
                }
            }
        },
        "github_com_alfascuf_PROD1_currency-service_internal_models.GetRateResponse": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "rate": {
                    "type": "number"
                },
                "target": {
                    "type": "string"
                }
            }
        }
    }
}